+++ packages/api/src/Domain/Client/Workspace/WorkspaceApiService.ts
@@ -1,3 +1,5 @@
+import { Uuid, WorkspaceType } from '@standardnotes/common'
+
 import { ErrorMessage } from '../../Error/ErrorMessage'
 import { ApiCallError } from '../../Error/ApiCallError'
 import { WorkspaceCreationResponse } from '../../Response/Workspace/WorkspaceCreationResponse'
@@ -6,6 +8,8 @@ import { WorkspaceServerInterface } from '../../Server/Workspace/WorkspaceServer
 import { WorkspaceApiServiceInterface } from './WorkspaceApiServiceInterface'
 import { WorkspaceApiOperations } from './WorkspaceApiOperations'
 
+import { WorkspaceInvitationResponse } from '../../Response'
+
 export class WorkspaceApiService implements WorkspaceApiServiceInterface {
   private operationsInProgress: Map<WorkspaceApiOperations, boolean>
 
@@ -13,10 +17,32 @@ export class WorkspaceApiService implements WorkspaceApiServiceInterface {
     this.operationsInProgress = new Map()
   }
 
+  async inviteToWorkspace(dto: { inviteeEmail: string; workspaceUuid: Uuid }): Promise<WorkspaceInvitationResponse> {
+    if (this.operationsInProgress.get(WorkspaceApiOperations.Inviting)) {
+      throw new ApiCallError(ErrorMessage.GenericInProgress)
+    }
+
+    this.operationsInProgress.set(WorkspaceApiOperations.Inviting, true)
+
+    try {
+      const response = await this.workspaceServer.inviteToWorkspace({
+        inviteeEmail: dto.inviteeEmail,
+        workspaceUuid: dto.workspaceUuid,
+      })
+
+      this.operationsInProgress.set(WorkspaceApiOperations.Inviting, false)
+
+      return response
+    } catch (error) {
+      throw new ApiCallError(ErrorMessage.GenericFail)
+    }
+  }
+
   async createWorkspace(dto: {
-    encryptedWorkspaceKey: string
-    encryptedPrivateKey: string
-    publicKey: string
+    workspaceType: WorkspaceType
+    encryptedWorkspaceKey?: string
+    encryptedPrivateKey?: string
+    publicKey?: string
     workspaceName?: string
   }): Promise<WorkspaceCreationResponse> {
     if (this.operationsInProgress.get(WorkspaceApiOperations.Creating)) {
@@ -27,6 +53,7 @@ export class WorkspaceApiService implements WorkspaceApiServiceInterface {
 
     try {
       const response = await this.workspaceServer.createWorkspace({
+        workspaceType: dto.workspaceType,
         encryptedPrivateKey: dto.encryptedPrivateKey,
         encryptedWorkspaceKey: dto.encryptedWorkspaceKey,
         publicKey: dto.publicKey,