+++ packages/api/src/Domain/Client/Workspace/WorkspaceApiServiceInterface.ts
@@ -1,10 +1,14 @@
-import { WorkspaceCreationResponse } from '../../Response'
+import { Uuid, WorkspaceType } from '@standardnotes/common'
+
+import { WorkspaceCreationResponse, WorkspaceInvitationResponse } from '../../Response'
 
 export interface WorkspaceApiServiceInterface {
   createWorkspace(dto: {
-    encryptedWorkspaceKey: string
-    encryptedPrivateKey: string
-    publicKey: string
+    workspaceType: WorkspaceType
+    encryptedWorkspaceKey?: string
+    encryptedPrivateKey?: string
+    publicKey?: string
     workspaceName?: string
   }): Promise<WorkspaceCreationResponse>
+  inviteToWorkspace(dto: { inviteeEmail: string; workspaceUuid: Uuid }): Promise<WorkspaceInvitationResponse>
 }