+++ packages/api/src/Domain/Server/Workspace/Paths.ts
@@ -1,5 +1,8 @@
+import { Uuid } from '@standardnotes/common'
+
 const WorkspacePaths = {
   createWorkspace: '/v1/workspaces',
+  inviteToWorkspace: (uuid: Uuid) => `/v1/workspaces/${uuid}/invites`,
 }
 
 export const Paths = {