+++ packages/api/src/Domain/Server/Workspace/WorkspaceServer.spec.ts
@@ -1,5 +1,7 @@
+import { WorkspaceType } from '@standardnotes/common'
 import { HttpServiceInterface } from '../../Http'
 import { WorkspaceCreationResponse } from '../../Response/Workspace/WorkspaceCreationResponse'
+import { WorkspaceInvitationResponse } from '../../Response/Workspace/WorkspaceInvitationResponse'
 
 import { WorkspaceServer } from './WorkspaceServer'
 
@@ -17,6 +19,7 @@ describe('WorkspaceServer', () => {
 
   it('should create a workspace', async () => {
     const response = await createServer().createWorkspace({
+      workspaceType: WorkspaceType.Private,
       encryptedPrivateKey: 'foo',
       encryptedWorkspaceKey: 'bar',
       publicKey: 'buzz',
@@ -25,6 +28,23 @@ describe('WorkspaceServer', () => {
     expect(response).toEqual({
       data: {
         uuid: '1-2-3',
+      },
+    })
+  })
+
+  it('should inivte to a workspace', async () => {
+    httpService.post = jest.fn().mockReturnValue({
+      data: { uuid: 'i-1-2-3' },
+    } as jest.Mocked<WorkspaceInvitationResponse>)
+
+    const response = await createServer().inviteToWorkspace({
+      inviteeEmail: 'test@test.te',
+      workspaceUuid: 'w-1-2-3',
+    })
+
+    expect(response).toEqual({
+      data: {
+        uuid: 'i-1-2-3',
       },
     })
   })