+++ packages/services/src/Domain/Workspace/WorkspaceClientInterface.ts
@@ -1,8 +1,12 @@
+import { Uuid, WorkspaceType } from '@standardnotes/common'
+
 export interface WorkspaceClientInterface {
   createWorkspace(dto: {
-    encryptedWorkspaceKey: string
-    encryptedPrivateKey: string
-    publicKey: string
+    workspaceType: WorkspaceType
+    encryptedWorkspaceKey?: string
+    encryptedPrivateKey?: string
+    publicKey?: string
     workspaceName?: string
   }): Promise<{ uuid: string } | null>
+  inviteToWorkspace(dto: { inviteeEmail: string; workspaceUuid: Uuid }): Promise<{ uuid: string } | null>
 }