+++ packages/services/src/Domain/Workspace/WorkspaceManager.ts
@@ -1,4 +1,6 @@
 import { WorkspaceApiServiceInterface } from '@standardnotes/api'
+import { Uuid, WorkspaceType } from '@standardnotes/common'
+
 import { InternalEventBusInterface } from '../Internal/InternalEventBusInterface'
 import { AbstractService } from '../Service/AbstractService'
 import { WorkspaceClientInterface } from './WorkspaceClientInterface'
@@ -11,10 +13,25 @@ export class WorkspaceManager extends AbstractService implements WorkspaceClient
     super(internalEventBus)
   }
 
+  async inviteToWorkspace(dto: { inviteeEmail: string; workspaceUuid: Uuid }): Promise<{ uuid: string } | null> {
+    try {
+      const result = await this.workspaceApiService.inviteToWorkspace(dto)
+
+      if (result.data.error !== undefined) {
+        return null
+      }
+
+      return result.data
+    } catch (error) {
+      return null
+    }
+  }
+
   async createWorkspace(dto: {
-    encryptedWorkspaceKey: string
-    encryptedPrivateKey: string
-    publicKey: string
+    workspaceType: WorkspaceType
+    encryptedWorkspaceKey?: string
+    encryptedPrivateKey?: string
+    publicKey?: string
     workspaceName?: string
   }): Promise<{ uuid: string } | null> {
     try {