+++ packages/releases/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.128](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.127...@standardnotes/releases@1.3.128) (2022-10-10)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.3.127](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.126...@standardnotes/releases@1.3.127) (2022-10-09)
+
+**Note:** Version bump only for package @standardnotes/releases
+
 ## [1.3.126](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.125...@standardnotes/releases@1.3.126) (2022-10-08)
 
 **Note:** Version bump only for package @standardnotes/releases